Adding pygame implementation & clean up.
