+++ requirements.txt
@@ -1,4 +1,3 @@
-arcade==2.6.17
 attrs==23.1.0
 cffi==1.16.0
 mypy==1.7.0
